refactor: invert sandboxes monitor runtime dependency
