+++ backend/sandboxes/resources/user_projection.py
@@ -6,7 +6,7 @@ from datetime import UTC, datetime
 from typing import Any
 
 import backend.sandboxes.resources.provider_boundary as resource_provider_boundary
-from backend.monitor.infrastructure.read_models import resource_runtime_service
+from backend.sandboxes.resources import runtime_service as resource_runtime_service
 from storage.models import map_sandbox_state_to_display_status
